Real-time clock: fix computation of RTC offset
The existing code that derives the RTC offset simply from the time
and date read from the RTC is taking for granted that the monotonic
clock starts counting from 0 when the VM is started. This is not
always the case: for example when running the VM under
KVM-accelerated Qemu, the monotonic clock returns the uptime of the
host machine; this results in the kernel computing a wall-clock
time with an offset given by the uptime of the host when the VM was
started.
The solution to this issue is to subtract from the initial RTC
value the initial monotonic timestamp when calculating the RTC
offset; this cannot be done in init_clock() but has to be done when
the platform clock is registered.