Priority queue: add pqueue_reorder() function
This function reorders the elements of a priority queue; it is
meant to be used when the prioriy value of queued elements may have
changed after the elements were inserted in the queue.
The pqueue unit tests now include code that calls pqueue_reorder().
This function will be used to reorder kernel timers following NTP
time updates.