mktime(): enforce POSIX compatibility of struct tm field
In the POSIX standard, the tm_mon field value ranges from 0 to 11,
and tm_year is the number of years since 1900.
This change makes mktime() conform to POSIX, so that a timestamp
broken-down via gmtime_r() can be re-converted to the original
number of seconds via mktime(). Handling of struct tm fields in
rtc_gettimeofday() is modified accordingly. This will make it
easier to compare timstamps written to the RTC with timestamps read
from it.